requirements.txt
fastapi
uvicorn
python-dotenv
transmission-rpc
